chore(js-sdk): refactor - structure
