+++ apps/js-sdk/firecrawl/src/utils/isZodSchema.ts
@@ -1,0 +1,5 @@
+import { ZodType } from "zod";
+
+export function isZodSchema(schema: unknown): schema is ZodType {
+  return schema instanceof ZodType;
+}
